Module Math

Some functors for various algebraic mathematical structures

author

San Vũ Ngọc, 2019-2020 IRMAR, Université de Rennes 1

Module types for algebraic structures

Groups, Rings, and Fields

module type Group = sig ... end

General groups, with multiplicative notation

module type AbelianGroup = sig ... end

Abelian groups, with additive notation

module type Rng = sig ... end

Rings without unit

module type Ring = sig ... end

Rings with unit

module type Field = sig ... end

Commutative fields

Modules and Vector Spaces

module type Module = sig ... end

Modules over a ring

module type VSpace = Module

Vector spaces over a field

Algebras and Lie Algebras

module type Algebra = sig ... end

Algebras over a field, with unit

module type LieAlg = sig ... end

Lie Algebras over a field

Usual number rings and fields

module Integers : Ring with type t = Z.t

Integers

module Int : Ring with type t = int

Machine integers

module Rationals : Field with type t = Q.t

Exact Rational numbers

module RealNumbers : Field with type t = float

Approximate real numbers (float)

module ComplexNumbers : sig ... end

Approximate complex numbers

module ComplexField : functor (F : Field) → sig ... end

Complexification of a field

module AlgebraOfField : functor (F : Field) → Algebra with type t = F.t and type scalar = F.t

View a Field as an Algebra over itself

Generic Multivariate Polynomials

Monomials

Monomials are the building bricks of Polynomials, but it is rarely necessary to access them directly.

module Iset : Set.S with type Iset.elt = int

Sets of integers

module type Monomial = sig ... end

Abelian group of abstract monomial in arbitrary number of coordinates.

module Monomial_generic : Monomial with type names = int → string option

Generic implementation of monomials

module Monomial1 : sig ... end

A simpler implementation for Monomials in 1 variable x_0

Polynomials

module type Polynomial = sig ... end

Polynomials over a ring.

module Polynomial_make : functor (M : Monomial) → functor (R : Ring) → Polynomial with type monomial = M.t and type scalar = R.t

The Polynomial functor

module Polynomial_generic : functor (R : Ring) → Polynomial with type monomial = Monomial_generic.t and type scalar = R.t

Generic polynomials with arbitrary number of variables

module RatPoly : Polynomial with type monomial = Monomial_generic.t and type scalar = Rationals.t

Polynomials with rational coefficients with arbitrary number of variables

module RealPoly : Polynomial with type monomial = Monomial_generic.t and type scalar = RealNumbers.t

Polynomials with real coefficients with arbitrary number of variables

Polynomials in one variable

module type Polynomial1 = sig ... end

Polynomials in one variable.

module Polynomial1_make : functor (R : Ring) → sig ... end

module RatPoly1 : Polynomial1 with type monomial = Monomial1.t and type scalar = Rationals.t and type generic = RatPoly.t

Polynomials in one variable with rational coefficients.

module RealPoly1 : Polynomial1 with type monomial = Monomial1.t and type scalar = RealNumbers.t and type generic = RealPoly.t

Polynomials in one variable with real (float) coefficients.