Module type Math.Algebra

Algebras over a field, with unit

A module of type Algebra can be used as a VSpace, or as a Ring.

type scalar

include Ring

include Rng

include AbelianGroup

type t

val zero : t

val is_zero : t → bool

val neg : t → t

val add : t → t → t

val (+) : t → t → t

val sub : t → t → t

val (-) : t → t → t

val mul : t → t → t

val (*) : t → t → t

val one : t

val mone : t

minus one

val of_int : int → t

type names

If the type names remains private, it means it has no use in the current module implementation.

val to_tex : ?⁠names:names → t → string

val of_tex : string → t

val of_scalar : scalar → t

of_scalar s is scal_mul s one

val scal_mul : scalar → t → t