Module type Math.Monomial

Abelian group of abstract monomial in arbitrary number of coordinates.

A monomial represents x_0^{a_0}··· x_n^{a_n} where x_j are formal indeterminates and a_j are integers (negative exponents are allowed). The group operation is the multiplication of monomials.

This group is isomorphic to the group of sequences of integers with finite support (the sequences of exponents a_0,a_1,...), where the group operation is now the addition. Hence the API exposes both the multiplicative and the additive notation.

include AbelianGroup

type t

val zero : t

val is_zero : t → bool

val neg : t → t

val add : t → t → t

val (+) : t → t → t

val sub : t → t → t

val (-) : t → t → t

val one : t

The constant mononial 1; same as zero, in additive notation.

val degree : t → int

val support : t → Iset.t

support m is the list of indices of m with non-zero exponents

val xi : int → t

xi i is the monomial "x_i"

val xin : int → int → t

xin i n is the monomial x_i^n

val mul : t → t → t

val (*) : t → t → t

multiplication of two monomials; same as addition of exponents.

val div : t → t → t

val (/) : t → t → t

div m1 m2 is the division of m1 by m2. This can produce negative exponents. Same as substraction of exponents.

val imul : int → t → t

imul i m is m muliplied by x_i

val of_list : (int * int) list → t

Create a monomial x_{i1}^{a1}x_{i2}^{a2}... from a list of the form [(i1,a1); (i2,a2), ...]. If the same index i appears several times, only the last one will be taken into account.

val to_list : t → (int * int) list

module Compare : sig ... end

type names

A variable of type names is used to print out the variables. The precise type depends on the implementation.

val to_tex : ?⁠names:names → t → string

val of_tex : string → t