Module type Math.VSpace

Vector spaces over a field

A module of type VSpace can be used as an AbelianGroup. VSpace is simply an alias to Module; it can be used when the scalar ring is a field.

type scalar

The scalar type should in principle include Rng.t, but this signature does not enforce it here.

include AbelianGroup

type t

val zero : t

Neutral element for addition.

val is_zero : t → bool

val neg : t → t

val add : t → t → t

val (+) : t → t → t

val sub : t → t → t

val (-) : t → t → t

val scal_mul : scalar → t → t